+++ common/src/main/java/pepjebs/mapatlases/config/MapAtlasesClientConfig.java
@@ -21,6 +21,9 @@ public class MapAtlasesClientConfig {
         ConfigBuilder builder = ConfigBuilder.create(MapAtlasesMod.MOD_ID, ConfigType.CLIENT);
 
         builder.push("minimap");
+        showsMapBackground = builder
+                .comment("Shows a background texture for each map, useful to see where each map ends")
+                .define("shows_map_background", false);
 
         hideWhenInHand = builder.comment("Hide minimap when holding the atlas in hand")
                 .define("hide_when_in_hand", true);
@@ -225,6 +228,7 @@ public class MapAtlasesClientConfig {
     public static final Supplier<Boolean> nightLightMap;
     public static final Supplier<Boolean> radarRotation;
     public static final Supplier<Boolean> convertXaero;
+    public static final Supplier<Boolean> showsMapBackground;
 
     public static final ModConfigHolder SPEC;